pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wilma-message-digest"
version = "0.1.0"
description = "Daily digest of Wilma school messages via Telegram"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
]

[project.scripts]
wilma-digest = "wilma_digest.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/wilma_digest"]
